production/requirements.txt
# Python 3.11+
# Core dependencies
openai-agents>=0.0.1
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Email (built-in, but documenting)
# smtplib, imaplib - Python standard library

# WhatsApp
requests>=2.31.0
# whapi-sdk  # Optional: use direct REST API instead

# Kafka
confluent-kafka>=2.3.0

# Sentiment Analysis
transformers>=4.36.0
torch>=2.1.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6  # For FastAPI form data
aiofiles>=23.2.1  # Async file operations

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0  # For FastAPI test client

# Logging & Monitoring
python-json-logger>=2.0.7
prometheus-client>=0.19.0

# Development
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
